Breadcrumb: Remove row div
It was causing the breadcrumb trail to look misaligned in most of the CDTS' themed templates. Only exception was the GCWeb content template - which never had a row div in the first place.

Bootstrap's row class is intended to be used as a container for a group of column grids (e.g. col-md-4). It basically negates left/right margins to allow a grid layout to use the full width. But WET/GCWeb's breadcrumbs don't use grids. So it isn't desirable to use the row class in that context.